backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
pinecone-client==3.0.2
anthropic==0.7.1
sentence-transformers==2.2.2
twilio==8.10.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
httpx==0.25.1
pydantic-settings==2.1.0
pydantic==2.5.0
alembic==1.12.1
cors==1.0.1
python-multipart==0.0.6
